feat: add authorization and account crud functions

+++ FoodVietAPI.Application/Utils/EnumUtil.cs
@@ -15,7 +15,7 @@ namespace CleanFoodVietAPI.Application.Utils
         }
 
         //Get enum name
-        public static string GetEnumName<T>(byte value)
+        public static string GetEnumName<T>(int value)
         {
             return Enum.GetName(typeof(T), value)!;
         }
